+++ pymatgen/analysis/interfaces/tests/test_substrate_analyzer.py
@@ -35,6 +35,9 @@ class SubstrateAnalyzerTest(PymatgenTest):
 
         matches = list(s.calculate(film, substrate, film_elac))
         self.assertEqual(len(matches), 192)
+        for match in matches:
+            assert match is not None
+            assert isinstance(match.match_area, float)
 
 
 if __name__ == "__main__":