+++ pymatgen/io/qchem/outputs.py
@@ -1247,8 +1247,8 @@ class QCOutput(MSONable):
         """
         dfs = nbo_parser(self.filename)
         nbo_data = {}
-        for key in dfs:
-            nbo_data[key] = [df.to_dict() for df in dfs[key]]
+        for key, value in dfs.items():
+            nbo_data[key] = [df.to_dict() for df in value]
         self.data["nbo_data"] = nbo_data
 
     def _check_completion_errors(self):